add error page, and improve routes
