+++ src/index.css
@@ -19,16 +19,3 @@ body {
   min-width: 100%;
   min-height: 100vh;
 }
-
-/* @media (prefers-color-scheme: light) {
-  :root {
-    color: #213547;
-    background-color: #ffffff;
-  }
-  a:hover {
-    color: #747bff;
-  }
-  button {
-    background-color: #f9f9f9;
-  }
-} */
