+++ src/routes/root.tsx
@@ -1,0 +1,12 @@
+import { Outlet } from "react-router-dom";
+import { LayoutPg } from "../components";
+
+export const Root = () => {
+  return (
+    <>
+      <LayoutPg />
+
+      <Outlet />
+    </>
+  );
+};
